+++ src/cjs/Triangle.d.ts
@@ -27,14 +27,16 @@
  * @modified  2025-14-16 Class `Triangle` now implements interface `IBounded`.
  * @modified  2025-14-16 Class `Triangle` now implements interface `Intersectable`.
  * @modified  2025-14-16 Added method `Triangle.move`.
- * @version   2.10.0
+ * @modified  2026-06-10 Refactoring the `Trianble.bounds` method and added a plain `Triangle.utils.bounds` method.
+ * @modified  2026-06-10 Refactoring the `Trianble.calcCircumCircle` method and added a plain `Triangle.utils.calcCircumCircle` method.
+ * @version   2.11.0
  *
  * @file Triangle
  * @fileoverview A simple triangle class: three vertices.
  * @public
  **/
 import { Bounds } from "./Bounds";
-import { Circle } from "./Circle";
+import { Circle, ICircle } from "./Circle";
 import { Line } from "./Line";
 import { Polygon } from "./Polygon";
 import { Vector } from "./Vector";
@@ -387,6 +389,20 @@ export declare class Triangle implements IBounded, SVGSerializable, Intersectabl
          * @param {XYCords} c - The first vertex.
          * @returns {nmber}
          */
-        determinant(a: XYCoords, b: XYCoords, c: XYCoords): number;
+        determinant: (a: XYCoords, b: XYCoords, c: XYCoords) => number;
+        bounds: (a: XYCoords, b: XYCoords, c: XYCoords) => Bounds;
+        /**
+         * Re-compute the circumcircle of this triangle (if the vertices
+         * have changed).
+         *
+         * The circumcenter and radius are stored in this.center and
+         * this.radius. There is a third result: radius_squared (for internal computations).
+         *
+         * @method calcCircumcircle
+         * @return void
+         * @instance
+         * @memberof Triangle
+         */
+        calcCircumcircle: (a: XYCoords, b: XYCoords, c: XYCoords) => ICircle;
     };
 }