+++ src/cjs/VertTuple.js
@@ -17,7 +17,8 @@
  * @modified 2025-03-31 Added the `VertTuple.revert` method.
  * @modified 2025-04-15 Changed param of `VertTuple.moveTo` method from `Vertex` to `XYCoords`.
  * @modified 2025-04-15 Added method `VertTuple.move` method.
- * @version 1.4.0
+ * @modified 2026-06-10 Adding helper function `VertTuple.utils.calcCircumcircle`.
+ * @version 1.5.0
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.VertTuple = void 0;
@@ -340,6 +341,10 @@ var VertTuple = /** @class */ (function () {
     VertTuple.vtutils = {
         dist2: function (v, w) {
             return (v.x - w.x) * (v.x - w.x) + (v.y - w.y) * (v.y - w.y);
+        },
+        calcCircumcircle: function (p1, p2) {
+            var p1x = p1.x, p1y = p1.y, p2x = p2.x, p2y = p2.y, cx = 0.5 * (p1x + p2x), cy = 0.5 * (p1y + p2y);
+            return { center: { x: cx, y: cy }, radius: Math.sqrt((p1x - cx) * (p1x - cx) + (p1y - cy) * (p1y - cy)) };
         }
     };
     return VertTuple;