+++ src/esm/Circle.js
@@ -14,13 +14,15 @@
  * @modified 2022-08-23 Added the `closestPoint` function.
  * @modified 2025-04-09 Added the `Circle.move(amount: XYCoords)` method.
  * @modified 2025-04-16 Class `Circle` now implements interface `Intersectable`.
- * @version  1.5.0
+ * @modified 2026-06-10 Adding the utility function `Circle.circleUtils.containsPoint`.
+ * @version  1.6.0
  **/
 import { Bounds } from "./Bounds";
 import { Line } from "./Line";
 import { UIDGenerator } from "./UIDGenerator";
 import { Vector } from "./Vector";
 import { Vertex } from "./Vertex";
+import { geomutils } from "./geomutils";
 /**
  * @classdesc A simple circle: center point and radius.
  *
@@ -326,6 +328,13 @@ Circle.circleUtils = {
         /* return new Vertex( Math.sin(angle) * radius,
                      Math.cos(angle) * radius ); */
         return new Vertex(Math.cos(angle) * radius, Math.sin(angle) * radius);
+    },
+    containsPoint: (point, circle) => {
+        // return (
+        //   (circle.center.x - point.x) * (circle.center.x - point.x) + (circle.center.y - point.y) * (circle.center.y - point.y) <=
+        //   circle.radius * circle.radius
+        // );
+        return geomutils.dist4(point.x, point.y, circle.center.x, circle.center.y) < circle.radius;
     }
 };
 //# sourceMappingURL=Circle.js.map