+++ src/ts/VertTuple.ts
@@ -16,12 +16,14 @@
  * @modified 2025-03-31 Added the `VertTuple.revert` method.
  * @modified 2025-04-15 Changed param of `VertTuple.moveTo` method from `Vertex` to `XYCoords`.
  * @modified 2025-04-15 Added method `VertTuple.move` method.
- * @version 1.4.0
+ * @modified 2026-06-10 Adding helper function `VertTuple.utils.calcCircumcircle`.
+ * @version 1.5.0
  */
 
 import { Vertex } from "./Vertex";
 import { UIDGenerator } from "./UIDGenerator";
 import { XYCoords, UID } from "./interfaces";
+import { ICircle } from "./Circle";
 
 /**
  * @classdesc An abstract base classes for vertex tuple constructs, like Lines or Vectors.
@@ -396,6 +398,17 @@ export class VertTuple<T extends VertTuple<T>> {
   static vtutils = {
     dist2: (v: XYCoords, w: XYCoords) => {
       return (v.x - w.x) * (v.x - w.x) + (v.y - w.y) * (v.y - w.y);
+    },
+
+    calcCircumcircle: (p1: XYCoords, p2: XYCoords): ICircle => {
+      const p1x = p1.x,
+        p1y = p1.y,
+        p2x = p2.x,
+        p2y = p2.y,
+        cx = 0.5 * (p1x + p2x),
+        cy = 0.5 * (p1y + p2y);
+
+      return { center: { x: cx, y: cy }, radius: Math.sqrt((p1x - cx) * (p1x - cx) + (p1y - cy) * (p1y - cy)) };
     }
   };
 }